Call \@@_store:nn before typesetting the mark in \@@_note:nn
For the sake of the multiple option, it is better if we have less events after
\@@_multiple_prepare:, not to affect \lastkern.  And \@@_store:nn, besides
performing a number of assignments, uses the postnotes/note/store hook.  It is
also arguably more correct placement to store the variables "right before" the
mark instead of "right after".  The downside is that we cannot capture changes
in variables occurring during the typesetting of the mark, but I don't see
much potential relevance to this.