Generate headers data with ltmarks (intermediary)
This commit is intermediary because we are just getting the data from ltmarks,
the UI is still kept untouched.  In other words, this commit keeps in record
that the \pnhd<info>first and \pnhd<info>last variables, which will be
deprecated soon, are being tested here and ensured to produce identical output
for the whole test suite (test logs vary only for \writes and \marks).  Note
that this result requires that we set checkruns=3 in build.lua given the
additional cross-reference needed for the ltmarks based variables.